Dégât sur zombie + crosshair
Possibilité de faire des dégâts sur les zombies et de les faires disparaitre en les tuants.
Ajout d'un crosshair vert.
Ajout des zombies rapides et des fonctionnalités de cette maj sur la scène principale.

+++ Assets/Scripts/Scriptable Objects/GunData.cs
@@ -5,19 +5,18 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "Gun", menuName = "Weapon/Gun")]
 public class GunData : ScriptableObject
 {
-
     [Header("Info")]
     public new string name;
 
     [Header("Shooting")]
-    public float Damage;
+    public int Damage;
     public float Max_Distance;
 
     [Header("Reloading")]
     public int Current_Ammo;
     public int Magazine_Size;
+
     [Tooltip("In RPM")] public float Fire_Rate;
     public float Reload_Time;
     [HideInInspector] public bool Reloading;
-
 }
